requirements.txt
mkdocs>=1.5.3
mkdocs-material>=9.5.0
pymdown-extensions>=10.7
pyyaml>=6.0

